Add support for multiple preferences and files
