+++ compiler/src/main/java/com/glassdoor/prefextensions/PreferenceUtils.kt
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2018 Shaishav Gandhi
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  You may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions
+ *  and limitations under the License.
+ */
+
+package com.glassdoor.prefextensions
+
+import com.squareup.kotlinpoet.ClassName
+import javax.lang.model.element.Element
+import javax.lang.model.element.ElementKind
+import javax.lang.model.element.PackageElement
+
+object PreferenceUtils {
+    /**
+     * Returns the top level classname of the given
+     * element.
+     *
+     * @param element
+     * @return [ClassName]
+     */
+    internal fun getClassName(element: Element): ClassName {
+        val classname = element.enclosingElement.simpleName.toString()
+        val packageName: String
+        var enclosing = element
+        while (enclosing.kind != ElementKind.PACKAGE) {
+            enclosing = enclosing.enclosingElement
+        }
+        val packageElement = enclosing as PackageElement
+        packageName = packageElement.toString()
+
+        return ClassName.bestGuess("$packageName.$classname")
+    }
+}
